requirements.txt
pyautogui==0.9.54
requests==2.32.5
pillow
praw
keyboard
discord.py
python-dotenv